+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix issue where using variables used within Not, Rnd(), Abs(), etc. in scripts were not being listed as parameters
 - Fix Variable Uses list to include variables used in script value calculations
 - Fix tilemap generation when number of unique tiles is above limits to match behaviour in v3.x
+- Fix issue where custom movement speed input would become focused unnecessarily
 
 ## [4.0.2] - 2024-07-29